+++ doc/changelog.rst
@@ -6,8 +6,10 @@ Version 0.7.4
 
 * metadata: add ``mem_max_rss`` and ``python_hash_seed``
 * add :func:`perf.python_implementation` and :func:`perf.python_has_jit`
-* TextRunner now recalibrates the benchmark if Python has a JIT compiler.
-  Add ``warmup_loops`` metadata in this case.
+* TextRunner now calibrates the benchmark while runnning warmup samples.
+  Add new warmup samples if needed until enough warmup raw samples are longer
+  than the minimum time.
+* Run warmups are now lists of (loops, raw_sample) rather than lists of samples
 * Fix "perf timeit" on PyPy
 
 Version 0.7.3 (2016-08-17)